fastapi_backend/test_login.py
import requests

url = "http://localhost:8000/api/v1/auth/login"
data = {"username": "admin", "password": "admin123"}

try:
    response = requests.post(url, json=data)
    print(f"Status Code: {response.status_code}")
    print(f"Response: {response.json()}")
except Exception as e:
    print(f"Error: {e}")
