fastapi_backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.7.1
pydantic-settings>=2.2.1
email-validator>=2.1.1
python-dotenv>=1.0.1
